requirements.txt
asgiref==3.5.2
coverage==6.4.1
crispy-bootstrap5==0.6
Django==4.0.5
django-crispy-forms==1.14.0
django-extensions==3.1.5
sqlparse==0.4.2
